Correctifs de Dijsktra et passage de Integer à Long

+++ PLD Agile 4IF/src/main/java/_4if/pld_agile_4if/models/CityMap.java
@@ -38,7 +38,7 @@ public class CityMap {
         return roadSegments;
     }
 
-    public ArrayList<RoadSegment> getOutGoingRoadSegments(int source) {
+    public ArrayList<RoadSegment> getOutGoingRoadSegments(long source) {
         ArrayList<RoadSegment> outgoing = new ArrayList<>();
         for (RoadSegment roadSegment : this.getRoadSegments()) {
             if ( roadSegment.getOrigin() == source ) {
